fix bug that spamcheck and viruscheck did not work
add additional error reporting to spam and viruscheck.
Use debug_fork option to activate

spam and viruscheck can now add a report to the header
of the message. See the spamcheck_include_report and
viruscheck_include_report options

additional name detection in adrbook.pl



git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@289 9f51437f-89f6-0310-a5ff-ff9a3806d2ee

+++ src/cgi-bin/openwebmail/modules/viruscheck.pl
@@ -2,17 +2,14 @@ package ow::viruscheck;
 #
 # viruscheck.pl - routines to call external checker to detect virus
 #
-# 2004/07/01 tung.AT.turtle.ee.ncku.edu.tw
-#
 
 use strict;
-use Fcntl qw(:DEFAULT :flock);
 require "modules/tool.pl";
 
 # sub scanmsg(pipecmd, msg_reference)
-# ret (0, null): message is clean
-#     (1, virusname): virus found
-#     (-1, runtime error): scanner has runtime error
+# ret (0, report): message is clean
+#     (1, report, virusname): virus found
+#     (-1, report): scanner has runtime error
 
 #
 # The following routines assume the clamav is used for the viruscheck.
@@ -24,61 +21,67 @@ require "modules/tool.pl";
 # output: stream: OK
 #         stream: VirusName FOUND
 sub scanmsg {
-   my $ret=pipecmd_msg(@_);
+   my $ret = pipecmd_msg(@_);
 
-   if ($ret=~/stream: OK/) {
-      return (0, '');
-   } elsif ($ret=~/stream: (.*?) FOUND/) {
-      return (1, $1);
+   my $report = $ret;
+   $report =~ s/[\n\r]+/ /gs;
+   $report =~ s/\s+$//gs;
+   $report =~ s/^\s+//gs;
+
+   if ($ret =~ m#^stream: OK#) {
+      return (0, $report);
+   } elsif ($ret =~ m#^stream: (.*?) FOUND#sg) {
+      my $virusname = $1;
+      return (1, $report, $virusname);
    } else {
-      $ret=~s/ERROR:\s+//;
-      return (-1, $ret);
+      return (-1, $report);
    }
 }
 
 # common routine, ret pipe output #########################################
 sub pipecmd_msg {
-   # refer to perlopentut and perlipc for more information.
-   my ($pipecmd, $r_message)=@_;
+   my ($pipecmd, $r_message) = @_;
 
-   my $username=getpwuid($>);	# username of euid
-   $pipecmd=~s/\@\@\@USERNAME\@\@\@/$username/g;
+   my $username = getpwuid($>); # username of euid
 
-   my ($outfh, $outfile)=ow::tool::mktmpfile('viruscheck.out');
-   my ($errfh, $errfile)=ow::tool::mktmpfile('viruscheck.err');
+   $pipecmd = ow::tool::untaint($pipecmd);
+   $pipecmd =~ s/\@\@\@USERNAME\@\@\@/$username/g;
 
-   local $|=1; # flush all output
+   my ($outfh, $outfile) = ow::tool::mktmpfile('viruscheck.out');
+   my ($errfh, $errfile) = ow::tool::mktmpfile('viruscheck.err');
 
-   # alias STDERR and STDOUT to get the output of the pipe
-   open(STDERR,">&=".fileno($errfh)) or return("dup STDERR failed: $!");
-   open(STDOUT,">&=".fileno($outfh)) or return("dup STDOUT failed: $!");
+   # STDIN gets closed if this is not a separate sub for some unknown reason
+   my $pipeerror = _pipecmd_msg($pipecmd, $r_message, $outfile, $errfile);
+   return $pipeerror if $pipeerror;
 
-   local $SIG{PIPE} = 'IGNORE'; # don't die if the fork pipe breaks
+   # slurp in all the output
+   local $/ = undef;
 
-   my ($out, $err, $errmsg);
-   open(P, "|$pipecmd") or return("can't fork to pipecmd: $! pipecmd: $pipecmd");
-   if (ref($r_message) eq 'ARRAY') {
-      print P @{$r_message} or return("can't write to pipe: $!");
-   } else {
-      print P ${$r_message} or return("can't write to pipe: $!");
-   }
-   close(P) or return("pipe broke - check connection to clamd - status: $?");
+   my $stderr=<$errfh>;
+   close($errfh) || return("could not close the stderrfile $errfile: $!");
 
-   close($errfh) or return("can't close errfh: $!");
-   close($outfh) or return("can't close outfh: $!");
+   my $stdout=<$outfh>;
+   close($outfh) || return("could not close the stdoutfile $outfile: $!");
 
-   sysopen(F, $errfile, O_RDONLY) or return("can't open errfile $errfile: $!");
-   $err = <F>;
-   close(F) or return("can't close errfile $errfile: $!");
    unlink $errfile;
-
-   sysopen(F, $outfile, O_RDONLY) or return("can't open outfile $outfile: $!");
-   $out = <F>;
-   close(F) or return("can't close outfile $outfile: $!");
    unlink $outfile;
 
-   return $err if (defined $err && $err ne '' && $err !~ m/^\s+$/s);
-   return $out;
+   foreach ($stderr, $stdout) {
+      return $_ if (defined $_ && $_ =~ m#\S#g);
+   }
+}
+
+sub _pipecmd_msg {
+   my ($pipecmd, $r_message, $outfile, $errfile) = @_;
+   open(P, "|$pipecmd 2>$errfile >$outfile") or return("pipecmd open failed: $!");
+   if (ref($r_message) eq 'ARRAY') {
+      print P @{$r_message} or return("print array to pipe failed: $!\n");
+   } else {
+      print P ${$r_message} or return("print string to pipe failed: $!\n");
+   }
+   # this close fails because clamdscan exits immediately, so do not check for error
+   close(P);
+   return 0;
 }
 
 1;
